added user management and some bugfixes

+++ eliminazione3.php
@@ -1,12 +1,22 @@
 <?php
-    $id = $_POST['id'];
-    $conn = new mysqli("localhost", "phpmadmin", "amministratore", "DBTelefonini");
-    if($conn->connect_errno)
+    session_start();
+    if(isset($_SESSION['id']))
     {
-        exit;
+        if($_SESSION['userlevel'] >= 2)
+        {
+            $id = $_POST['id'];
+            $conn = new mysqli("localhost", "phpmadmin", "amministratore", "DBTelefonini");
+            if($conn->connect_errno)
+            {
+                exit;
+            }
+            $query = 'DELETE FROM telefoni WHERE id = '.$id.';';
+            $result = $conn->query($query);
+            $conn->close();
+            HEADER("location:index.php");
+        }
+    } else
+    {
+        header("Location: index.php");
     }
-    $query = 'DELETE FROM telefoni WHERE id = '.$id.';';
-    $result = $conn->query($query);
-	$conn->close();
-    HEADER("location:index.php");
 ?>
